consolidate table order components
