+++ context/app/static/js/components/search/Results/style.ts
@@ -3,8 +3,6 @@ import Table from '@mui/material/Table';
 import TableRow from '@mui/material/TableRow';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
-import ArrowUpward from '@mui/icons-material/ArrowUpwardRounded';
-import ArrowDownward from '@mui/icons-material/ArrowDownwardRounded';
 
 import { HeaderCell } from 'js/shared-styles/tables';
 
@@ -70,31 +68,8 @@ const StyledTableCell = styled(TableCell)({
   },
 });
 
-const sharedArrowStyles = {
-  verticalAlign: 'text-top',
-  fontSize: '1.1rem',
-} as const;
-
-const ArrowUpOn = styled(ArrowUpward)(sharedArrowStyles);
-
-const ArrowDownOn = styled(ArrowDownward)(sharedArrowStyles);
-
-const ArrowDownOff = styled(ArrowDownward)({
-  ...sharedArrowStyles,
-  opacity: '40%',
-});
-
 const StyledHeaderCell = styled(HeaderCell)({
   whiteSpace: 'nowrap',
 });
 
-export {
-  StyledTable,
-  StyledTableRow,
-  StyledTableBody,
-  StyledTableCell,
-  ArrowUpOn,
-  ArrowDownOn,
-  ArrowDownOff,
-  StyledHeaderCell,
-};
+export { StyledTable, StyledTableRow, StyledTableBody, StyledTableCell, StyledHeaderCell };
