+++ context/app/static/js/shared-styles/tables/TableOrdering/TableOrdering.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { ArrowDownOff, ArrowDownOn, ArrowUpOn } from './style';
+
+export type SortDirection = 'asc' | 'desc';
+
+function OrderIcon({ direction, isCurrentSortField }: { direction: SortDirection; isCurrentSortField: boolean }) {
+  if (!isCurrentSortField) return <ArrowDownOff />;
+  if (direction === 'asc') return <ArrowUpOn />;
+  if (direction === 'desc') return <ArrowDownOn />;
+}
+
+function getSortOrder({ direction, isCurrentSortField }: { direction: SortDirection; isCurrentSortField: boolean }) {
+  if (!isCurrentSortField) {
+    return 'desc';
+  }
+
+  return direction === 'desc' ? 'asc' : 'desc';
+}
+
+export { OrderIcon, getSortOrder };
