+++ context/app/static/js/shared-styles/tables/TableOrdering/index.ts
@@ -1,0 +1,3 @@
+import { OrderIcon, getSortOrder } from './TableOrdering';
+
+export { OrderIcon, getSortOrder };
